+++ src/main/java/roomescape/reservation/dto/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.dto.request;
 
 import java.time.LocalDate;
 
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 
 import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotNull;
-import roomescape.auth.dto.LoginMember;
 
 public record ReservationRequest(
         @FutureOrPresent @JsonFormat(pattern = "yyyy-MM-dd") LocalDate date,
